feat: create models
Implement models for users (admin & instructors), courses and its corresponding lectures. The lecture model has date and it also references the instructor to whom the lecture of a particular course is assigned to. The user model has a role property which identifies whether the logged in user is an ADMIN or INSTRUCTOR